Updated text Inputs
