+++ src/components/forms-inputs/text-input-form.component.tsx
@@ -1,8 +1,6 @@
-import { SIZES, useThemeColor } from "@/src/constants/theme";
-import { useState } from "react";
-import { StyleSheet, TextInput, View } from "react-native";
+import { View } from "react-native";
 import { Controller, Control } from "react-hook-form";
-import ThemeText from "../theme-text.component";
+import ThemeTextInput from "./theme-text-input.component";
 
 export default function TextInputForm({
   name,
@@ -10,60 +8,46 @@ export default function TextInputForm({
   placeHolder,
   label,
   type,
+  onChange,
+  onBlur,
+  value,
 }: {
   name: string;
   label?: string;
   type: string;
   placeHolder?: string;
-  control: Control;
+  control?: Control;
+  onChange?: () => void;
+  onBlur?: () => void;
+  value?: string;
 }) {
-  const COLORS = useThemeColor();
-  const [borderColor, setBorderColor] = useState(COLORS.borderColor);
   return (
     <View>
-      {label && <ThemeText style={styles.labelText}>{label}</ThemeText>}
-      <Controller
-        name={name}
-        control={control}
-        render={({ field: { onChange, onBlur, value } }) => (
-          <TextInput
-            onFocus={() => {
-              setBorderColor(COLORS.primary);
-            }}
-            onBlur={() => {
-              setBorderColor(COLORS.borderColor);
-              onBlur();
-            }}
-            onChangeText={onChange}
-            value={value}
-            placeholder={placeHolder}
-            placeholderTextColor={COLORS.borderColor}
-            selectionColor={COLORS.primary}
-            autoCapitalize={type === "email" ? "none" : "sentences"}
-            style={[
-              styles.textInput,
-              {
-                borderColor,
-                color: COLORS.text,
-              },
-            ]}
-          />
-        )}
-      />
+      {control ? (
+        <Controller
+          name={name}
+          control={control}
+          render={({ field: { onChange, onBlur, value } }) => (
+            <ThemeTextInput
+              onChange={onChange}
+              onBlur={onBlur}
+              value={value}
+              placeHolder={placeHolder}
+              type={type}
+              label={label}
+            />
+          )}
+        />
+      ) : (
+        <ThemeTextInput
+          onChange={onChange}
+          onBlur={onBlur}
+          value={value}
+          placeHolder={placeHolder}
+          type={type}
+          label={label}
+        />
+      )}
     </View>
   );
 }
-
-const styles = StyleSheet.create({
-  textInput: {
-    width: "100%",
-    height: 50,
-    borderRadius: 20,
-    borderWidth: 1,
-    paddingHorizontal: SIZES.marginOrPadding.large,
-  },
-  labelText: {
-    fontSize: SIZES.fontSize.medium,
-    marginBottom: SIZES.marginOrPadding.small,
-  },
-});
